+++ draft-packages/filter-menu-button/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.4.2](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-filter-menu-button@2.4.1...@kaizen/draft-filter-menu-button@2.4.2) (2022-05-03)
+
+**Note:** Version bump only for package @kaizen/draft-filter-menu-button
+
+
+
+
+
 ## [2.4.1](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-filter-menu-button@2.4.0...@kaizen/draft-filter-menu-button@2.4.1) (2022-05-02)
 
 **Note:** Version bump only for package @kaizen/draft-filter-menu-button