+++ draft-packages/hero-card/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.4](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-hero-card@2.0.3...@kaizen/draft-hero-card@2.0.4) (2022-05-03)
+
+**Note:** Version bump only for package @kaizen/draft-hero-card
+
+
+
+
+
 ## [2.0.3](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-hero-card@2.0.2...@kaizen/draft-hero-card@2.0.3) (2022-04-29)
 
 **Note:** Version bump only for package @kaizen/draft-hero-card